decouple ParametersMergeTest from CheckerTolerantYamlFileLoaderTest
